backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
